backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.6.0
python-decouple==3.8
bcrypt==4.2.1
psycopg2-binary==2.9.10